Make the dropzone replicatable
Add a button to make the dropzone replicatable. It opens a window
for the user to resize the dropzone before replicating it. Needs
its own window, because the main window size might not be shrinkable
enough for the desired replicant size.

When files are dropped, forward refs by launching the Filer app, so
it'll still work when replicated.

Removed localization preparations, because that'll only work with a
layout-aware RulesEditWindow