Bugfix and live settings for folders and rules
Major bug fixed: The RuleEditWindow didn't inform the RuleTab that
a rule was added, edited or removed.

Bug fixed: Dropping files resulted in ProcessFiles() being run twice. Now
we run it only once in ReadyToRun().

Update the button to start/stop AutoFiler on tab creation. Eliminates a
possible delay of 0.5s until the BMessageRunner comes around for the first
time.

The AutoFiler folder settings and the Filer rules are now saved and applied
live and don't wait for the window to close. The only setting saved on close
is now the Autofiler autorun checkbox.