Refined method of selecting a file in Tracker
We now send two messages to Tracker to see if our file is
selected. Depending on the Tracker setting, the window title
may be the only the folder name or the whole folder path.

To not flood Tracker, only send the messages every 0.1 sec.

After our file is confirmed to be selected, the Tracker window may
still be poplated with new files, which would scroll our file out
of the visible area. Wait another 0.2 sec and send the selection
message again. May not be enough for really big folders or slow
media, but a longer delay would result in responsiveness issues.