Fix crashes
There were two double frees: BLoopers destroy themselves before returning.
So after  window->Go(doAll);  and  app->Run();  the window / app are
already freed.

The EditRuleWindow crashes when it tries to send a message and the
fCaller already went away. Using a messenger and SendMessage() does
the checking for us.

Thanks to DeadYak, for guiding me on the bug hunt.

Fixes #111